Refactor release_controller.process module
