+++ apiserver/paasng/paas_wl/platform/external/local.py
@@ -17,35 +17,12 @@ to the current version of the project delivered to anyone in the future.
 """
 from typing import Dict, Optional
 
-from paasng.engine.constants import JobStatus
-from paasng.engine.deploy.release import ApplicationReleaseMgr
-from paasng.engine.models import Deployment
-from paasng.engine.models.offline import OfflineOperation
 from paasng.platform.applications.models import ModuleEnvironment
-from paasng.platform.applications.signals import module_environment_offline_success
 from paasng.platform.operations.models import Operation
 
 
 class LocalPlatformSvcClient:
     """Client for "apiserver" module, uses local module"""
-
-    def finish_release(self, deployment_id: str, status: str, error_detail: str):
-        mgr = ApplicationReleaseMgr.from_deployment_id(deployment_id)
-        mgr.callback_release(JobStatus(status), error_detail)
-
-    def finish_archive(self, operation_id: str, status: str, error_detail: str):
-        offline_op = OfflineOperation.objects.get(id=operation_id)
-        if status == JobStatus.SUCCESSFUL:
-            offline_op.set_successful()
-        else:
-            offline_op.set_failed(error_detail)
-
-        module_environment_offline_success.send(
-            sender=OfflineOperation, offline_instance=offline_op, environment=offline_op.app_environment.environment
-        )
-
-    def retrieve_deployment(self, deployment_id: str) -> Deployment:
-        return Deployment.objects.get(pk=deployment_id)
 
     def create_operation_log(
         self,
