+++ apiserver/paasng/paasng/engine/deploy/release.py
@@ -19,22 +19,29 @@ to the current version of the project delivered to anyone in the future.
 """Releasing process of an application deployment
 """
 import logging
+import typing
 from typing import Optional
 
+from blue_krill.async_utils.poll_task import CallbackHandler, CallbackResult, CallbackStatus, TaskPoller
 from django.utils.translation import gettext as _
+from pydantic import ValidationError as PyDanticValidationError
 
 from paasng.engine.configurations.building import get_processes_by_build
 from paasng.engine.configurations.config_var import get_env_variables
 from paasng.engine.configurations.image import update_image_runtime_config
 from paasng.engine.configurations.ingress import AppDefaultDomains, AppDefaultSubpaths
-from paasng.engine.constants import JobStatus
+from paasng.engine.constants import JobStatus, ReleaseStatus
 from paasng.engine.deploy.engine_svc import EngineDeployClient
 from paasng.engine.models.deployment import Deployment
 from paasng.engine.models.phases import DeployPhaseTypes
 from paasng.engine.models.processes import ProcessManager
+from paasng.engine.processes.wait import AbortedDetails, wait_for_release
 from paasng.engine.signals import on_release_created
 from paasng.engine.workflow import DeployStep
 from paasng.platform.applications.models import ModuleEnvironment
+
+logger = logging.getLogger(__name__)
+
 
 logger = logging.getLogger(__name__)
 
@@ -91,10 +98,61 @@ class ApplicationReleaseMgr(DeployStep):
         self.state_mgr.finish(status, err_detail=error_detail, write_to_stream=True)
 
 
+class ReleaseResultHandler(CallbackHandler):
+    """Result handler for `wait_for_release`"""
+
+    def handle(self, result: CallbackResult, poller: TaskPoller):
+        is_interrupted, error_detail = self.get_error_detail(result)
+        # Transform release_status
+        if is_interrupted:
+            status = ReleaseStatus.INTERRUPTED
+        elif error_detail:
+            status = ReleaseStatus.FAILED
+        else:
+            status = ReleaseStatus.SUCCESSFUL
+
+        deployment_id = poller.params['extra_params']['deployment_id']
+        self.finish_release(str(deployment_id), status, error_detail)
+
+    def finish_release(self, deployment_id: str, status: ReleaseStatus, error_detail: str):
+        """Finish the release"""
+        mgr = ApplicationReleaseMgr.from_deployment_id(deployment_id)
+        # TODO: transfer ReleaseStatus to JobStatus
+        mgr.callback_release(JobStatus(status.value), error_detail)
+
+    def get_error_detail(self, result: CallbackResult) -> typing.Tuple[bool, str]:
+        """Get detailed error message. if error message was empty, release was considered succeeded
+
+        :returns: (is_interrupted, error_msg)
+        """
+        if result.status == CallbackStatus.TIMEOUT:
+            return False, "Timeout: polling release's status taking too long to complete"
+
+        if result.status == CallbackStatus.NORMAL:
+            aborted_details = self.get_aborted_details(result)
+            if not (aborted_details and aborted_details.aborted):
+                return False, ''
+
+            assert aborted_details.policy is not None, 'policy must not be None'  # Make type checker happy
+            reason = aborted_details.policy.reason
+            return aborted_details.policy.is_interrupted, f'Release aborted, reason: {reason}'
+
+        return False, 'Release failed, internal error'
+
+    def get_aborted_details(self, result: CallbackResult) -> Optional[AbortedDetails]:
+        """If current release was aborted, return detailed info"""
+        try:
+            details = AbortedDetails.parse_obj(result.data)
+        except PyDanticValidationError:
+            return None
+        return details
+
+
 def create_release(env: ModuleEnvironment, build_id: str, deployment: Optional[Deployment] = None) -> str:
     """Create a new release by calling engine's API
 
     :param deployment: if not given, will try using the latest succeed deployment for getting desc env vars
+    :return: The ID of the created release object.
     """
     if not deployment:
         try:
@@ -113,4 +171,14 @@ def create_release(env: ModuleEnvironment, build_id: str, deployment: Optional[D
         deployment_id = None
 
     extra_envs = get_env_variables(env, deployment=deployment)
-    return EngineDeployClient(env.get_engine_app()).create_release(build_id, deployment_id, extra_envs, procfile)
+
+    # Create the release and start the background task to wait for the release if needed
+    release = EngineDeployClient(env.get_engine_app()).create_release(build_id, deployment_id, extra_envs, procfile)
+    if deployment_id:
+        wait_for_release(
+            env=env,
+            release_version=release.version,
+            result_handler=ReleaseResultHandler,
+            extra_params={"deployment_id": deployment_id},
+        )
+    return str(release.uuid)
