+++ apiserver/paasng/paasng/engine/processes/__init__.py
@@ -16,3 +16,5 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
+# import handlers module to set up handler functions
+from . import handlers  # noqa
