+++ apiserver/paasng/paasng/engine/processes/wait.py
@@ -26,41 +26,43 @@ import django.dispatch
 from blue_krill.async_utils.poll_task import PollingMetadata, PollingResult, PollingStatus, TaskPoller
 from pydantic import BaseModel, validator
 
-from paas_wl.platform.applications.models import WlApp
-from paas_wl.platform.external.client import get_local_plat_client
-from paas_wl.release_controller.process.events import ProcEventsProducer
-from paas_wl.release_controller.process.models import PlainProcess, condense_processes
-from paas_wl.release_controller.process.utils import ProcessesSnapshotStore
 from paas_wl.workloads.processes.controllers import get_processes_status
+from paasng.engine.models import Deployment
+from paasng.engine.processes.events import ProcEventsProducer
+from paasng.engine.processes.models import PlainProcess, condense_processes
+from paasng.engine.processes.utils import ProcessesSnapshotStore
+from paasng.platform.applications.models import ModuleEnvironment
 
 logger = logging.getLogger(__name__)
 
 processes_updated = django.dispatch.Signal(providing_args=['events', 'extra_params'])
 
 
-def wait_for_all_stopped(wl_app: WlApp, result_handler: Type, extra_params: Optional[Dict] = None):
+def wait_for_all_stopped(env: ModuleEnvironment, result_handler: Type, extra_params: Optional[Dict] = None):
     """Wait for processes to be fully stopped
 
-    :param wl_app: WlApp object
+    :param env: ModuleEnvironment object
     :param result_handler: type to handle poll result
     :param extra_params: extra params, use it to provide data for result handler
     """
     extra_params = extra_params or {}
-    params = {'wl_app_id': wl_app.pk, 'extra_params': extra_params}
+    params = {'env_id': env.pk, 'extra_params': extra_params}
     WaitForAllStopped.start(params, result_handler)
 
 
-def wait_for_release(wl_app: WlApp, release_version: int, result_handler: Type, extra_params: Optional[Dict] = None):
+def wait_for_release(
+    env: ModuleEnvironment, release_version: int, result_handler: Type, extra_params: Optional[Dict] = None
+):
     """Wait for processes to be updated to given release version, includes all process instances
 
-    :param wl_app: WlApp object
+    :param env: ModuleEnvironment object
     :param release_version: release version, included in each instance
     :param result_handler: type to handle poll result
     :param extra_params: extra params, use it to provide data for result handler
     """
     extra_params = extra_params or {}
     params = {
-        'wl_app_id': wl_app.pk,
+        'env_id': env.pk,
         'broadcast_enabled': True,
         'release_version': release_version,
         'extra_params': extra_params,
@@ -110,18 +112,18 @@ class WaitProcedurePoller(TaskPoller):
 
     def __init__(self, params: Dict, metadata: PollingMetadata):
         super().__init__(params, metadata)
-        self.wl_app = WlApp.objects.get(uuid=self.params['wl_app_id'])
+        self.env = ModuleEnvironment.objects.get(pk=self.params['env_id'])
 
         self.broadcast_enabled = bool(self.params.get('broadcast_enabled'))
         self.extra_params = self.params.get('extra_params', {})
-        self.store = ProcessesSnapshotStore(self.wl_app)
+        self.store = ProcessesSnapshotStore(self.env)
 
     def query(self) -> PollingResult:
         """Start polling query"""
         current_processes = self._get_current_processes()
 
         already_waited = time.time() - self.metadata.query_started_at
-        logger.info(f'wait procedure started {already_waited} seconds, wl_app: {self.wl_app.name}')
+        logger.info(f'wait procedure started {already_waited} seconds, env: {self.env}')
         # Check all abort policies
         for policy in self.abort_policies:
             if policy.evaluate(current_processes, already_waited, self.extra_params):
@@ -162,11 +164,16 @@ class WaitProcedurePoller(TaskPoller):
 
     def _get_current_processes(self) -> List[PlainProcess]:
         """Get current process list"""
-        return condense_processes(get_processes_status(self.wl_app))
+        wl_app = self.env.wl_app
+        return condense_processes(get_processes_status(wl_app))
 
     def _get_last_processes(self) -> Optional[List[PlainProcess]]:
         """Get process list of last polling action"""
-        return self.store.get()
+        try:
+            return self.store.get()
+        except Exception as e:
+            logger.warning('Failed to get last processes, error: %s', e)
+            return None
 
     def get_status(self, processes: List[PlainProcess]) -> PollingResult:
         raise NotImplementedError()
@@ -227,8 +234,8 @@ class UserInterruptedPolicy(AbortPolicy):
             return False
 
         try:
-            deployment = get_local_plat_client().retrieve_deployment(deployment_id=deployment_id)
-        except Exception:
+            deployment = Deployment.objects.get(pk=deployment_id)
+        except Deployment.DoesNotExist:
             logger.warning('Deployment not exists for UserInterruptedPolicy, will not proceed.')
             return False
         return bool(deployment.release_int_requested_at)
@@ -247,7 +254,7 @@ class WaitForAllStopped(WaitProcedurePoller):
                 logger.info(f'Process {process.type} still have {count} instances')
                 return PollingResult.doing()
 
-        logger.info(f'No instances found, all processes has been stopped for wl-app: {self.wl_app.name}')
+        logger.info(f'No instances found, all processes has been stopped for env: {self.env}')
         return PollingResult.done()
 
 
@@ -265,12 +272,10 @@ class WaitForReleaseAllReady(WaitProcedurePoller):
         """Check if all where processes was updated to given release_version"""
         for process in processes:
             if not process.is_all_ready(self.release_version):
-                logger.info(
-                    f'Process {process.type} was not updated to {self.release_version}, ' f'wl-app: {self.wl_app.name}'
-                )
+                logger.info(f'Process {process.type} was not updated to {self.release_version}, env: {self.env}')
                 return PollingResult.doing()
 
-        logger.info(f'All processes has been updated to {self.release_version}, wl-app: {self.wl_app.name}')
+        logger.info(f'All processes has been updated to {self.release_version}, env: {self.env}')
         return PollingResult.done()
